% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{detect_kmers_across_sequences}
\alias{detect_kmers_across_sequences}
\title{Count k-mers across sequences.}
\usage{
detect_kmers_across_sequences(sequences, kmer_size)
}
\description{
Count k-mers across sequences.
}
