% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{kmers}
\alias{kmers}
\title{Temporary function that returns the indexes (in base4) of the kemers.}
\usage{
kmers(sequence, k)
}
\description{
Temporary function that returns the indexes (in base4) of the kemers.
}
